Fixed: returning correct error when anime is not found
Previously, there was a bug that led the server to return INTERNAL_SERVER_ERROR when anime was not found, instead of returning the correct 404 NO_ANIME error. It was because Mongoose tried to hydrate an unexisting anime from cache, and so it threw a different internal error that wasn't a not found one. I fixed it by retrieving the cached anime, checking if it exists and only THEN hydrate and return it, otherwise return an error.