Implementing aggregation for getAnime in AnimeService
This is useful because now any request that needs an anime, including getting an anime by id (for the client or for operating with vote/follow endpoint), will have aggregated data such as scores and followers informations. This was mandatory for clients that needed such data without having to call the fetch endpoint multiple times.